Develop Version V-0.3.0

+++ database/migrations/2023_12_29_161907_create_aviones_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('modelo');
             $table->string('fabricante');
             $table->double('precio');
-            $table->unsignedBigInteger('capacidad');
+            $table->unsignedInteger('capacidad');
             $table->unsignedInteger('rango');
             $table->string('img');
             $table->unsignedFloat('costePorKm');
